Rename main component files to component name
* and use index.js just for easy includes
